Moved all the style to css as class moved all the logic from index to App removed use of ref
