+++ mylearning-app/src/index.css
@@ -17,4 +17,43 @@ code {
   gap: 10px;
 }
 
+.counter {
+  display: flex;
+  justify-content: center;
+  align-items: center;
+  font-size: 24px;
+  padding-top: 20px;
+}
+
+.button {
+  padding: 10px 15px;
+  font-size: 16px;
+  background-color : #f0f0f0;
+  color : #000;
+  font-weight : normal
+}
+
+.selected-button
+{
+  padding: 10px 15px;
+  font-size: 16px;
+  background-color : #007bff;
+  color : #fff;
+  font-weight : bold
+}
+.searchbox {
+    padding: 5px;
+    font-size: 16px;
+    margin-bottom: 10px;
+    height: 20px;
+  }
+
+  .container {
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+    gap: 20px;
+    margin-top: 50px;
+  }
+
