+++ mylearning-app/src/index.js
@@ -2,46 +2,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import './index.css';
-import Counter from './components/Counter';
-import SearchBox from './components/SearchForm';
 import reportWebVitals from './reportWebVitals';
-import MovieGenres from './components/GenreSelect';
+
+import App from './App';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
 
-let counterValue = 10;
-
-
-function handleSearch(query) {
-  console.log("Search triggered with query:", query);
-}
-
-let selectedGenre = "ALL"; 
-const genresList = ["ALL","DOCUMENTARY", "COMEDY", "HORROR", "CRIME"]; 
-function handleGenreSelect(newGenre) {
-  selectedGenre = newGenre; 
-  console.log("New selected movie:", newGenre);
-}
-
 root.render(
   <React.StrictMode>
-    <div
-      style={{
-        display: "flex",
-        flexDirection: "column",
-        alignItems: "center",
-        gap: "20px",
-        marginTop: "50px",
-      }}
-    >
-      <SearchBox initialQuery="Hello" onSearch={handleSearch} />
-      <Counter initialValue={counterValue} />
-      <MovieGenres
-        genres={genresList}
-        selectedGenre={selectedGenre}
-        onSelect={handleGenreSelect}
-      />
-    </div>
+    <App></App>
   </React.StrictMode>
 );
